+++ savedsearch-web/src/main/java/oracle/sysman/emSDK/emaas/platform/savedsearch/ws/rest/search/SearchAPI.java
@@ -2,6 +2,7 @@ package oracle.sysman.emSDK.emaas.platform.savedsearch.ws.rest.search;
 
 import java.math.BigInteger;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import javax.ws.rs.Consumes;
@@ -26,6 +27,7 @@ import oracle.sysman.SDKImpl.emaas.platform.savedsearch.util.IdGenerator;
 import oracle.sysman.SDKImpl.emaas.platform.savedsearch.util.LogUtil;
 import oracle.sysman.SDKImpl.emaas.platform.savedsearch.util.RegistryLookupUtil;
 import oracle.sysman.SDKImpl.emaas.platform.savedsearch.util.ZDTContext;
+import oracle.sysman.emSDK.emaas.platform.savedsearch.exception.EMAnalyticsDatabaseUnavailException;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.exception.EMAnalyticsFwkException;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.model.Category;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.model.CategoryManager;
@@ -35,10 +37,15 @@ import oracle.sysman.emSDK.emaas.platform.savedsearch.model.Search;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.model.SearchManager;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.model.SearchParameter;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.model.TenantContext;
+import oracle.sysman.emSDK.emaas.platform.savedsearch.restnotify.WidgetDeletionNotification;
+import oracle.sysman.emSDK.emaas.platform.savedsearch.restnotify.WidgetNotificationType;
+import oracle.sysman.emSDK.emaas.platform.savedsearch.restnotify.WidgetNotifyEntity;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.ws.exception.EMAnalyticsWSException;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.ws.rest.util.JsonUtil;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.ws.rest.util.StringUtil;
 import oracle.sysman.emSDK.emaas.platform.savedsearch.ws.rest.util.ValidationUtil;
+import oracle.sysman.emaas.platform.savedsearch.entity.EmAnalyticsSearch;
+import oracle.sysman.emaas.platform.savedsearch.services.DependencyStatus;
 import oracle.sysman.emaas.platform.savedsearch.targetmodel.services.OdsDataService;
 import oracle.sysman.emaas.platform.savedsearch.targetmodel.services.OdsDataServiceImpl;
 
@@ -234,6 +241,9 @@ public class SearchAPI
 		int statusCode = 201;
 		SearchManager sman = SearchManager.getInstance();
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			Search searchObj = createSearchObjectForAdd(inputJsonObj);
 			Search savedSearch = sman.saveSearch(searchObj);
 			
@@ -325,8 +335,17 @@ public class SearchAPI
 		SearchManager sman = SearchManager.getInstance();
 		
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			odsService.deleteOdsEntity(searchId);
-			sman.deleteSearch(searchId, false);
+			EmAnalyticsSearch eas = sman.deleteSearch(searchId, false);
+			// TODO: when merging with ZDT, this deletionTime should be from the APIGW request
+			Date deletionTime = DateUtil.getCurrentUTCTime();
+			WidgetNotifyEntity wne = new WidgetNotifyEntity(eas, deletionTime, WidgetNotificationType.DELETE);
+			if (eas.getIsWidget() == 1L) {
+				new WidgetDeletionNotification().notify(wne);
+			}
 		}
 		catch (EMAnalyticsFwkException e) {
 			LOGGER.error(e.getLocalizedMessage());
@@ -334,6 +353,32 @@ public class SearchAPI
 		}
 
 		return Response.status(statusCode).build();
+	}
+
+	@DELETE
+	@Produces(MediaType.APPLICATION_JSON)
+	public Response deleteSearchByName(@QueryParam("searchName") String name, @QueryParam("isExactly") String isExactly) {
+		LogUtil.getInteractionLogger().info("Service calling to (DELETE) /savedsearch/v1/search?searchName={}&isExactly={}", name, isExactly);
+		SearchManager searchManager = SearchManager.getInstance();
+		if (isExactly == null) {
+			LOGGER.debug("isExactly is null and set the default value true");
+			isExactly = "true";
+		}
+		if (!"true".equalsIgnoreCase(isExactly) && !"false".equalsIgnoreCase(isExactly)) {
+			LOGGER.error("The param isExactly is invalid");
+			return Response.status(Response.Status.BAD_REQUEST).entity("The param isExactly is invalid").build();
+		}
+		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
+			LOGGER.debug("Calling searchManager.deleteSearchByName");
+			searchManager.deleteSearchByName(name, Boolean.valueOf(isExactly));
+		} catch (EMAnalyticsFwkException e) {
+			LOGGER.error(e.getLocalizedMessage());
+			return Response.status(e.getStatusCode()).entity(e.getMessage()).build();
+		}
+		return Response.status(Response.Status.OK).build();
 	}
 
 	/**
@@ -454,6 +499,9 @@ public class SearchAPI
 		Search searchObj;
 		SearchManager sman = SearchManager.getInstance();
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			if (updateCategory != null && "ORACLE_INTERNAL".equals(updateCategory)) {
 				searchObj = createSearchObjectForEdit(inputJsonObj, sman.getSearch(searchId), true);
 
@@ -490,6 +538,9 @@ public class SearchAPI
 		Search savedSearch = null;
 		
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			savedSearch = sman.getSearch(searchId);
 		} catch (EMAnalyticsFwkException e) {
 			message = e.getMessage();
@@ -548,6 +599,9 @@ public class SearchAPI
 		// find and return the whole object
 		SearchManager sman = SearchManager.getInstance();
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			Search searchObj = sman.getSearchWithoutOwner(searchId);
 			jsonObj = EntityJsonUtil.getFullSearchJsonObj(uri.getBaseUri(), searchObj);
 			message = jsonObj.toString();
@@ -705,7 +759,9 @@ public class SearchAPI
 		int statusCode = 200;
 		SearchManager sman = SearchManager.getInstance();
 		try {
-
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			Search searchObj = sman.getSearchWithoutOwner(searchid);
 			String[] pathArray = null;
 			if (bPath) {
@@ -724,12 +780,61 @@ public class SearchAPI
 
 		return Response.status(statusCode).entity(message).build();
 	}
+	
+	@PUT
+	@Path("/list")
+	@Consumes(MediaType.APPLICATION_JSON)
+	@Produces(MediaType.APPLICATION_JSON)
+	public Response getSearchList(JSONArray inputJsonArray)
+	{
+		LogUtil.getInteractionLogger().info("Service calling to (PUT) /savedsearch/v1/search/list");
+		List<BigInteger> searchIdList = new ArrayList<BigInteger>();
+		try {
+			for(int i = 0; i < inputJsonArray.length(); i++) {
+				searchIdList.add(new BigInteger(inputJsonArray.getString(i)));
+			}
+		} catch(NullPointerException npe) {
+			return Response.status(404).entity("search id list is empty").build();
+		} catch(JSONException je) {
+			return Response.status(404).entity("invalid search id in search id list").build();
+		} catch(NumberFormatException nfe) {
+			return Response.status(404).entity("invalid search id in search id list").build();
+		}
+		
+		String message = null;
+		int statusCode = 200;
+		JSONArray outputJsonArray = new JSONArray();
+		SearchManager sman = SearchManager.getInstance();
+		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
+			sman.updateLastAccessDate(searchIdList);
+			
+			List<Search> searchList = sman.getSearchListByIds(searchIdList);
+			for(Search searchObj : searchList) {
+				outputJsonArray.put(EntityJsonUtil.getFullSearchJsonObj(uri.getBaseUri(), searchObj));
+			}
+			message = outputJsonArray.toString();
+		} catch (EMAnalyticsFwkException e) {
+			LOGGER.error(e.getLocalizedMessage());
+			statusCode = e.getStatusCode();
+			message = e.getMessage();
+			LOGGER.error((TenantContext.getContext() != null ? TenantContext.getContext().toString() : "") + e.getMessage(),
+					e.getStatusCode());
+		}
+
+		return Response.status(statusCode).entity(message).build();
+	}
 
 	private Response updateLastAccessTime(BigInteger searchId)
 	{
 		String message = null;
 		int statusCode = 200;
 		try {
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			SearchManager sman = SearchManager.getInstance();
 			java.util.Date date = sman.modifyLastAccessDate(searchId);
 			message = String.valueOf(DateUtil.getDateFormatter().format(date));
@@ -737,7 +842,7 @@ public class SearchAPI
 		catch (EMAnalyticsFwkException e) {
 			LOGGER.error(e.getLocalizedMessage());
 			message = e.getMessage();
-			statusCode = 404;
+			statusCode = e.getStatusCode();
 		}
 		return Response.status(statusCode).entity(message).build();
 	}
@@ -753,6 +858,9 @@ public class SearchAPI
 		String version;
 		String rel;
 		try{
+			if (!DependencyStatus.getInstance().isDatabaseUp()) {
+				throw new EMAnalyticsDatabaseUnavailException();
+			}
 			Search search = searchManager.getSearch(searchid);
 			if(search.getCategoryId()==null){
 				return Response.status(Response.Status.NOT_FOUND).entity("Search with id "+searchid
@@ -781,7 +889,7 @@ public class SearchAPI
 		}catch (EMAnalyticsFwkException e) {
 			LOGGER.error((TenantContext.getContext() != null ? TenantContext.getContext().toString() : "") + e.getMessage(),
 					e.getStatusCode());
-			return Response.status(Response.Status.NOT_FOUND).entity(e.getMessage()).build();
+			return Response.status(e.getStatusCode()).entity(e.getMessage()).build();
 		}
 	}
 	private Search createSearchObjectForAdd(JSONObject json) throws EMAnalyticsWSException
